Put mock classes in their own file, test_mocks.py
We want to decouple the dependency on test_windows.py of other test files.
