+++ mypy.ini
@@ -41,6 +41,10 @@ disallow_untyped_defs = False
 check_untyped_defs = True
 disallow_untyped_defs = False
 
+[mypy-plugin.core.test_mocks]
+check_untyped_defs = True
+disallow_untyped_defs = False
+
 [mypy-plugin.core.test_session]
 check_untyped_defs = True
 disallow_untyped_defs = False
