+++ plugin/core/test_session.py
@@ -1,71 +1,18 @@
-from .types import ClientConfig, LanguageConfig, ClientStates, Settings
 from .sessions import create_session, Session
-from .protocol import Request, Notification
-from .logging import debug
-
+from .test_mocks import MockClient
+from .test_mocks import TEST_CONFIG
+from .test_mocks import TEST_LANGUAGE
+from .types import ClientConfig
+from .types import ClientStates
+from .types import Settings
 import unittest
 import unittest.mock
+
 try:
     from typing import Any, List, Dict, Tuple, Callable, Optional
     assert Any and List and Dict and Tuple and Callable and Optional and Session
 except ImportError:
     pass
-
-
-basic_responses = {
-    'initialize': {
-        'capabilities': {
-            'testing': True,
-            'hoverProvider': True,
-            'completionProvider': {
-                'triggerCharacters': ['.'],
-                'resolveProvider': False
-            },
-            'textDocumentSync': True,
-            'definitionProvider': True,
-            'typeDefinitionProvider': True,
-            'declarationProvider': True,
-            'implementationProvider': True
-        }
-    }
-}
-
-
-class MockClient():
-    def __init__(self, async_response=None) -> None:
-        self.responses = basic_responses
-        self._notifications = []  # type: List[Notification]
-        self._async_response_callback = async_response
-
-    def send_request(self, request: Request, on_success: 'Callable', on_error: 'Callable' = None) -> None:
-        response = self.responses.get(request.method)
-        debug("TEST: responding to", request.method, "with", response)
-        if self._async_response_callback:
-            self._async_response_callback(lambda: on_success(response))
-        else:
-            on_success(response)
-
-    def send_notification(self, notification: Notification) -> None:
-        self._notifications.append(notification)
-
-    def on_notification(self, name, handler: 'Callable') -> None:
-        pass
-
-    def on_request(self, name, handler: 'Callable') -> None:
-        pass
-
-    def set_error_display_handler(self, handler: 'Callable') -> None:
-        pass
-
-    def set_crash_handler(self, handler: 'Callable') -> None:
-        pass
-
-    def exit(self) -> None:
-        pass
-
-
-test_language = LanguageConfig("test", ["source.test"], ["Plain Text"])
-test_config = ClientConfig("test", [], None, languages=[test_language])
 
 
 class SessionTest(unittest.TestCase):
@@ -77,7 +24,7 @@ class SessionTest(unittest.TestCase):
     # @unittest.skip("need an example config")
     def test_can_create_session(self):
 
-        config = ClientConfig("test", ["ls"], None, [], [], None, [test_language])
+        config = ClientConfig("test", ["ls"], None, [], [], None, [TEST_LANGUAGE])
         project_path = "/"
         session = self.assert_if_none(
             create_session(config, project_path, dict(), Settings()))
@@ -91,7 +38,7 @@ class SessionTest(unittest.TestCase):
         project_path = "/"
         post_initialize_callback = unittest.mock.Mock()
         session = self.assert_if_none(
-            create_session(config=test_config,
+            create_session(config=TEST_CONFIG,
                            project_path=project_path,
                            env=dict(),
                            settings=Settings(),
@@ -109,7 +56,7 @@ class SessionTest(unittest.TestCase):
         pre_initialize_callback = unittest.mock.Mock()
         post_initialize_callback = unittest.mock.Mock()
         session = self.assert_if_none(
-            create_session(config=test_config,
+            create_session(config=TEST_CONFIG,
                            project_path=project_path,
                            env=dict(),
                            settings=Settings(),
@@ -129,7 +76,7 @@ class SessionTest(unittest.TestCase):
         post_initialize_callback = unittest.mock.Mock()
         post_exit_callback = unittest.mock.Mock()
         session = self.assert_if_none(
-            create_session(config=test_config,
+            create_session(config=TEST_CONFIG,
                            project_path=project_path,
                            env=dict(),
                            settings=Settings(),
