+++ plugin/core/test_sublime.py
@@ -1,5 +1,3 @@
-
-
 DIALOG_CANCEL = 0
 DIALOG_YES = 1
 DIALOG_NO = 2
