+++ plugin/core/test_windows.py
@@ -1,13 +1,22 @@
-from .windows import WindowManager, WindowRegistry, ViewLike
-from .diagnostics import WindowDiagnostics
-from .sessions import create_session, Session
-from .test_session import MockClient, test_config, test_language
-from .test_rpc import MockSettings
-from .events import global_events
-from .types import ClientConfig, LanguageConfig
 from . import test_sublime as test_sublime
-# from .logging import set_debug_logging, debug
-import os
+from .diagnostics import WindowDiagnostics
+from .events import global_events
+from .sessions import create_session
+from .sessions import Session
+from .test_mocks import MockClient
+from .test_mocks import MockConfigs
+from .test_mocks import MockDocuments
+from .test_mocks import MockHandlerDispatcher
+from .test_mocks import MockSettings
+from .test_mocks import MockView
+from .test_mocks import MockWindow
+from .test_mocks import TEST_CONFIG
+from .test_mocks import TestDocumentHandlerFactory
+from .test_mocks import TestGlobalConfigs
+from .types import ClientConfig
+from .types import LanguageConfig
+from .windows import WindowManager
+from .windows import WindowRegistry
 import tempfile
 import unittest
 
@@ -19,210 +28,6 @@ except ImportError:
     pass
 
 
-class MockSublimeSettings(object):
-    def __init__(self, values):
-        self._values = values
-
-    def get(self, key, default=None):
-        return self._values.get(key, default)
-
-    def set(self, key, value):
-        self._values[key] = value
-
-
-class MockView(object):
-    def __init__(self, file_name):
-        self._file_name = file_name
-        self._window = None
-        self._settings = MockSublimeSettings({"syntax": "Plain Text"})
-        self._status = dict()  # type: Dict[str, str]
-        self._text = "asdf"
-
-    def file_name(self):
-        return self._file_name
-
-    def set_window(self, window):
-        self._window = window
-
-    def set_status(self, key, status):
-        self._status[key] = status
-
-    def window(self):
-        return self._window
-
-    def settings(self):
-        return self._settings
-
-    def substr(self, region):
-        return self._text
-
-    def size(self):
-        return len(self._text)
-
-    def sel(self):
-        return [test_sublime.Region(1, 1)]
-
-    def score_selector(self, region, scope: str) -> int:
-        return 1
-
-    def buffer_id(self):
-        return 1
-
-
-class MockHandlerDispatcher(object):
-    def __init__(self, can_start: bool = True) -> None:
-        self._can_start = can_start
-        self._initialized = set()  # type: Set[str]
-
-    def on_start(self, config_name: str, window) -> bool:
-        return self._can_start
-
-    def on_initialized(self, config_name: str, window, client):
-        self._initialized.add(config_name)
-
-
-class MockWindow(object):
-    def __init__(self, files_in_groups: 'List[List[ViewLike]]' = []) -> None:
-        self._files_in_groups = files_in_groups
-        self._is_valid = True
-        self._folders = [os.path.dirname(__file__)]
-        self._default_view = MockView(None)
-        self._project_data = None  # type: Optional[Dict[str, Any]]
-        self.commands = []  # type: List[Tuple[str, Dict[str, Any]]]
-
-    def id(self):
-        return 0
-
-    def folders(self):
-        return self._folders
-
-    def set_folders(self, folders):
-        self._folders = folders
-
-    def num_groups(self):
-        return len(self._files_in_groups)
-
-    def active_group(self):
-        return 0
-
-    def project_data(self) -> Optional[dict]:
-        return self._project_data
-
-    def set_project_data(self, data: Optional[dict]):
-        self._project_data = data
-
-    def active_view(self) -> Optional[ViewLike]:
-        return self.active_view_in_group(0)
-
-    def close(self):
-        self._is_valid = False
-
-    def is_valid(self):
-        return self._is_valid
-
-    def extract_variables(self):
-        return {
-            "project_path": os.path.dirname(__file__)
-        }
-
-    def active_view_in_group(self, group):
-        if group < len(self._files_in_groups):
-            files = self._files_in_groups[group]
-            if len(files) > 0:
-                return files[0]
-            else:
-                return self._default_view
-
-    def add_view_in_group(self, group, view):
-        self._files_in_groups[group].append(view)
-
-    def status_message(self, msg: str) -> None:
-        pass
-
-    def views(self):
-        views = []
-        for views_in_group in self._files_in_groups:
-            if len(views_in_group) < 1:
-                views.append(self._default_view)
-            else:
-                for view in views_in_group:
-                    views.append(view)
-        return views
-
-    def find_open_file(self, path: str) -> 'Optional[ViewLike]':
-        pass
-
-    def run_command(self, command_name: str, command_args: 'Dict[str, Any]') -> None:
-        self.commands.append((command_name, command_args))
-
-
-class TestGlobalConfigs(object):
-    def for_window(self, window):
-        return MockConfigs()
-
-
-class MockConfigs(object):
-    def __init__(self):
-        self.all = [test_config]
-
-    def is_supported(self, view):
-        return any(self.scope_configs(view))
-
-    def scope_configs(self, view, point=None):
-        if view.file_name() is None:
-            return [None]
-        else:
-            return [test_config]
-
-    def syntax_configs(self, view):
-        if view.settings().get("syntax") == "Plain Text":
-            return [test_config]
-        else:
-            return []
-
-    def syntax_supported(self, view: ViewLike) -> bool:
-        return view.settings().get("syntax") == "Plain Text"
-
-    def syntax_config_languages(self, view: ViewLike) -> 'Dict[str, LanguageConfig]':
-        if self.syntax_supported(view):
-            return {
-                "test": test_language
-            }
-        else:
-            return {}
-
-    def update(self, configs: 'List[ClientConfig]') -> None:
-        pass
-
-    def disable(self, config_name: str) -> None:
-        pass
-
-
-class MockDocuments(object):
-    def __init__(self):
-        self._documents = []  # type: List[str]
-        self._sessions = {}  # type: Dict[str, Session]
-
-    def add_session(self, session: 'Session') -> None:
-        self._sessions[session.config.name] = session
-
-    def remove_session(self, config_name: str) -> None:
-        del self._sessions[config_name]
-
-    def handle_view_opened(self, view: ViewLike):
-        file_name = view.file_name()
-        if file_name:
-            self._documents.append(file_name)
-
-    def reset(self):
-        self._documents = []
-
-
-class TestDocumentHandlerFactory(object):
-    def for_window(self, window, configs):
-        return MockDocuments()
-
-
 def mock_start_session(window: MockWindow,
                        project_path: str,
                        config: ClientConfig,
@@ -230,7 +35,7 @@ def mock_start_session(window: MockWindow,
                        on_post_initialize: 'Callable[[Session], None]',
                        on_post_exit: 'Callable[[str], None]') -> 'Optional[Session]':
     return create_session(
-        config=test_config,
+        config=TEST_CONFIG,
         project_path=project_path,
         env=dict(),
         settings=MockSettings(),
@@ -278,7 +83,7 @@ class WindowManagerTests(unittest.TestCase):
         wm.start_active_views()
 
         # session must be started (todo: verify session is ready)
-        self.assertIsNotNone(wm.get_session(test_config.name))
+        self.assertIsNotNone(wm.get_session(TEST_CONFIG.name))
         self.assertListEqual(docs._documents, [__file__])
 
     def test_can_open_supported_view(self):
@@ -288,14 +93,14 @@ class WindowManagerTests(unittest.TestCase):
                            MockHandlerDispatcher())
 
         wm.start_active_views()
-        self.assertIsNone(wm.get_session(test_config.name))
+        self.assertIsNone(wm.get_session(TEST_CONFIG.name))
         self.assertListEqual(docs._documents, [])
 
         # session must be started (todo: verify session is ready)
         view = MockView(__file__)
 
         wm.activate_view(view)
-        self.assertIsNotNone(wm.get_session(test_config.name))
+        self.assertIsNotNone(wm.get_session(TEST_CONFIG.name))
         self.assertEqual(len(docs._sessions), 1)
 
     def test_can_restart_sessions(self):
@@ -305,7 +110,7 @@ class WindowManagerTests(unittest.TestCase):
         wm.start_active_views()
 
         # session must be started (todo: verify session is ready)
-        self.assertIsNotNone(wm.get_session(test_config.name))
+        self.assertIsNotNone(wm.get_session(TEST_CONFIG.name))
 
         # our starting document must be loaded
         self.assertListEqual(docs._documents, [__file__])
@@ -313,7 +118,7 @@ class WindowManagerTests(unittest.TestCase):
         wm.restart_sessions()
 
         # session must be started (todo: verify session is ready)
-        self.assertIsNotNone(wm.get_session(test_config.name))
+        self.assertIsNotNone(wm.get_session(TEST_CONFIG.name))
 
         # our starting document must be loaded
         self.assertListEqual(docs._documents, [__file__])
@@ -327,7 +132,7 @@ class WindowManagerTests(unittest.TestCase):
         wm.start_active_views()
 
         # session must be started (todo: verify session is ready)
-        self.assertIsNotNone(wm.get_session(test_config.name))
+        self.assertIsNotNone(wm.get_session(TEST_CONFIG.name))
 
         # our starting document must be loaded
         self.assertListEqual(docs._documents, [__file__])
@@ -348,7 +153,7 @@ class WindowManagerTests(unittest.TestCase):
         wm.start_active_views()
 
         # session must be started (todo: verify session is ready)
-        self.assertIsNotNone(wm.get_session(test_config.name))
+        self.assertIsNotNone(wm.get_session(TEST_CONFIG.name))
 
         # our starting document must be loaded
         self.assertListEqual(docs._documents, [__file__])
@@ -375,15 +180,15 @@ class WindowManagerTests(unittest.TestCase):
         wm.start_active_views()
 
         # session must be started (todo: verify session is ready)
-        self.assertIsNotNone(wm.get_session(test_config.name))
+        self.assertIsNotNone(wm.get_session(TEST_CONFIG.name))
 
         # our starting document must be loaded
         self.assertListEqual(docs._documents, [__file__])
 
-        wm._handle_server_crash(test_config)
+        wm._handle_server_crash(TEST_CONFIG)
 
         # session must be started (todo: verify session is ready)
-        self.assertIsNotNone(wm.get_session(test_config.name))
+        self.assertIsNotNone(wm.get_session(TEST_CONFIG.name))
 
         # our starting document must be loaded
         self.assertListEqual(docs._documents, [__file__])
@@ -397,10 +202,10 @@ class WindowManagerTests(unittest.TestCase):
         wm.start_active_views()
 
         # session must be started (todo: verify session is ready)
-        self.assertIsNotNone(wm.get_session(test_config.name))
+        self.assertIsNotNone(wm.get_session(TEST_CONFIG.name))
 
         # our starting document must be loaded
         self.assertListEqual(docs._documents, [__file__])
 
         # client_start_listeners, client_initialization_listeners,
-        self.assertTrue(test_config.name in dispatcher._initialized)
+        self.assertTrue(TEST_CONFIG.name in dispatcher._initialized)
